requirements.txt
Flask==2.1.1
sentence-transformers==2.5.0
faiss-cpu==1.7.1
pandas==1.3.4
numpy==1.21.2
pickle5==0.0.7
dotenv==0.19.2
flask-cors==3.0.10
langchain-core==0.1.0
